implemented encoder and adapted decoder for zoneFiles
