+++ utils/testUtil/testUtil.go
@@ -1,0 +1,387 @@
+package testUtil
+
+import (
+	"net"
+	"rains/rainslib"
+	"testing"
+)
+
+func CheckMessage(m1, m2 rainslib.RainsMessage, t *testing.T) {
+	if m1.Token != m2.Token {
+		t.Error("Token mismatch")
+	}
+	if len(m1.Capabilities) != len(m2.Capabilities) {
+		t.Error("Capabilities mismatch")
+	}
+	for i := 0; i < len(m1.Capabilities); i++ {
+		if m1.Capabilities[i] != m2.Capabilities[i] {
+			t.Error("Capabilities mismatch")
+		}
+	}
+	CheckSignatures(m1.Signatures, m2.Signatures, t)
+	if len(m1.Content) != len(m2.Content) {
+		t.Error("Message Content length mismatch")
+	}
+	for i, s1 := range m1.Content {
+		switch s1 := s1.(type) {
+		case *rainslib.AssertionSection:
+			if s2, ok := m2.Content[i].(*rainslib.AssertionSection); ok {
+				CheckAssertion(s1, s2, t)
+				continue
+			}
+			t.Errorf("Types at position %d of Content slice are different", i)
+		case *rainslib.ShardSection:
+			if s2, ok := m2.Content[i].(*rainslib.ShardSection); ok {
+				CheckShard(s1, s2, t)
+				continue
+			}
+			t.Errorf("Types at position %d of Content slice are different", i)
+		case *rainslib.ZoneSection:
+			if s2, ok := m2.Content[i].(*rainslib.ZoneSection); ok {
+				CheckZone(s1, s2, t)
+				continue
+			}
+			t.Errorf("Types at position %d of Content slice are different", i)
+		case *rainslib.QuerySection:
+			if s2, ok := m2.Content[i].(*rainslib.QuerySection); ok {
+				CheckQuery(s1, s2, t)
+				continue
+			}
+			t.Errorf("Types at position %d of Content slice are different", i)
+		case *rainslib.NotificationSection:
+			if s2, ok := m2.Content[i].(*rainslib.NotificationSection); ok {
+				CheckNotification(s1, s2, t)
+				continue
+			}
+			t.Errorf("Types at position %d of Content slice are different", i)
+		case *rainslib.AddressAssertionSection:
+			if s2, ok := m2.Content[i].(*rainslib.AddressAssertionSection); ok {
+				CheckAddressAssertion(s1, s2, t)
+				continue
+			}
+			t.Errorf("Types at position %d of Content slice are different", i)
+		case *rainslib.AddressZoneSection:
+			if s2, ok := m2.Content[i].(*rainslib.AddressZoneSection); ok {
+				CheckAddressZone(s1, s2, t)
+				continue
+			}
+			t.Errorf("Types at position %d of Content slice are different", i)
+		case *rainslib.AddressQuerySection:
+			if s2, ok := m2.Content[i].(*rainslib.AddressQuerySection); ok {
+				CheckAddressQuery(s1, s2, t)
+				continue
+			}
+			t.Errorf("Types at position %d of Content slice are different", i)
+		default:
+			t.Errorf("Unsupported section type: %T", s1)
+		}
+	}
+}
+
+func CheckSignatures(s1, s2 []rainslib.Signature, t *testing.T) {
+	if len(s1) != len(s2) {
+		t.Error("Signature count mismatch")
+	}
+	for i := 0; i < len(s1); i++ {
+		if s1[i].Algorithm != s2[i].Algorithm {
+			t.Errorf("Signature algorithm mismatch in %d. Signature", i)
+		}
+		if s1[i].KeySpace != s2[i].KeySpace {
+			t.Errorf("Signature KeySpace mismatch in %d. Signature", i)
+		}
+		if s1[i].ValidSince != s2[i].ValidSince {
+			t.Errorf("Signature ValidSince mismatch in %d. Signature", i)
+		}
+		if s1[i].ValidUntil != s2[i].ValidUntil {
+			t.Errorf("Signature ValidUntil mismatch in %d. Signature", i)
+		}
+		switch s1[i].Algorithm {
+		case rainslib.Ed25519:
+			d1 := s1[i].Data.([]byte)
+			d2 := s2[i].Data.([]byte)
+			if len(d1) != len(d2) {
+				t.Errorf("Signature data length mismatch in %d. Signature", i)
+			}
+			for j := 0; j < len(d1); j++ {
+				if d1[j] != d2[j] {
+					t.Errorf("Signature data mismatch at %d. byte in %d. Signature", j, i)
+				}
+			}
+		}
+	}
+}
+
+func CheckAssertion(a1, a2 *rainslib.AssertionSection, t *testing.T) {
+	if a1.Context != a2.Context {
+		t.Error("Assertion Context mismatch")
+	}
+	if a1.SubjectZone != a2.SubjectZone {
+		t.Error("Assertion SubjectZone mismatch")
+	}
+	if a1.SubjectName != a2.SubjectName {
+		t.Error("Assertion SubjectName mismatch")
+	}
+	CheckSignatures(a1.Signatures, a2.Signatures, t)
+	CheckObjects(a1.Content, a2.Content, t)
+}
+
+func CheckShard(s1, s2 *rainslib.ShardSection, t *testing.T) {
+	if s1.Context != s2.Context {
+		t.Error("Shard context mismatch")
+	}
+	if s1.SubjectZone != s2.SubjectZone {
+		t.Error("Shard subjectZone mismatch")
+	}
+	if s1.RangeFrom != s2.RangeFrom {
+		t.Error("Shard RangeFrom mismatch")
+	}
+	if s1.RangeTo != s2.RangeTo {
+		t.Error("Shard RangeTo mismatch")
+	}
+	CheckSignatures(s1.Signatures, s2.Signatures, t)
+	if len(s1.Content) != len(s2.Content) {
+		t.Error("Shard Content length mismatch")
+	}
+	for i, a1 := range s1.Content {
+		CheckAssertion(a1, s2.Content[i], t)
+	}
+}
+
+func CheckZone(z1, z2 *rainslib.ZoneSection, t *testing.T) {
+	if z1.Context != z2.Context {
+		t.Error("Zone context mismatch")
+	}
+	if z1.SubjectZone != z2.SubjectZone {
+		t.Error("Zone subjectZone mismatch")
+	}
+	CheckSignatures(z1.Signatures, z2.Signatures, t)
+	if len(z1.Content) != len(z2.Content) {
+		t.Error("Zone Content length mismatch")
+	}
+	for i, s1 := range z1.Content {
+		switch s1 := s1.(type) {
+		case *rainslib.AssertionSection:
+			if s2, ok := z2.Content[i].(*rainslib.AssertionSection); ok {
+				CheckAssertion(s1, s2, t)
+				continue
+			}
+			t.Errorf("Types at position %d of Content slice are different", i)
+		case *rainslib.ShardSection:
+			if s2, ok := z2.Content[i].(*rainslib.ShardSection); ok {
+				CheckShard(s1, s2, t)
+				continue
+			}
+			t.Errorf("Types at position %d of Content slice are different", i)
+		default:
+			t.Errorf("Unsupported section type: %T", s1)
+		}
+	}
+}
+
+func CheckQuery(q1, q2 *rainslib.QuerySection, t *testing.T) {
+	if q1.Context != q2.Context {
+		t.Error("Query context mismatch")
+	}
+	if q1.Expires != q2.Expires {
+		t.Error("Query Expires mismatch")
+	}
+	if q1.Name != q2.Name {
+		t.Error("Query Name mismatch")
+	}
+	if q1.Token != q2.Token {
+		t.Error("Query Token mismatch")
+	}
+	if q1.Type != q2.Type {
+		t.Error("Query Type mismatch")
+	}
+	if len(q1.Options) != len(q2.Options) {
+		t.Error("Query Option length mismatch")
+	}
+	for i, o1 := range q1.Options {
+		if o1 != q2.Options[i] {
+			t.Errorf("Query Option at position %d mismatch", i)
+		}
+	}
+}
+
+func CheckNotification(n1, n2 *rainslib.NotificationSection, t *testing.T) {
+	if n1.Type != n2.Type {
+		t.Error("Notification Type mismatch")
+	}
+	if n1.Token != n2.Token {
+		t.Error("Notification Token mismatch")
+	}
+	if n1.Data != n2.Data {
+		t.Error("Notification Data mismatch")
+	}
+}
+
+func CheckAddressAssertion(a1, a2 *rainslib.AddressAssertionSection, t *testing.T) {
+	if a1.Context != a2.Context {
+		t.Error("AddressAssertion Context mismatch")
+	}
+	CheckSignatures(a1.Signatures, a2.Signatures, t)
+	CheckSubjectAddress(a1.SubjectAddr, a2.SubjectAddr, t)
+	CheckObjects(a1.Content, a2.Content, t)
+}
+
+func CheckAddressZone(z1, z2 *rainslib.AddressZoneSection, t *testing.T) {
+	if z1.Context != z2.Context {
+		t.Error("AddressZone Context mismatch")
+	}
+	CheckSignatures(z1.Signatures, z2.Signatures, t)
+	CheckSubjectAddress(z1.SubjectAddr, z2.SubjectAddr, t)
+	if len(z1.Content) != len(z2.Content) {
+		t.Error("AddressZone Content length mismatch")
+	}
+	for i, a1 := range z1.Content {
+		CheckAddressAssertion(a1, z2.Content[i], t)
+	}
+}
+
+func CheckAddressQuery(q1, q2 *rainslib.AddressQuerySection, t *testing.T) {
+	if q1.Context != q2.Context {
+		t.Error("AddressQuery context mismatch")
+	}
+	if q1.Expires != q2.Expires {
+		t.Error("AddressQuery Expires mismatch")
+	}
+	if q1.Token != q2.Token {
+		t.Error("AddressQuery Token mismatch")
+	}
+	if q1.Types != q2.Types {
+		t.Error("AddressQuery Type mismatch")
+	}
+	if len(q1.Options) != len(q2.Options) {
+		t.Error("AddressQuery Option length mismatch")
+	}
+	for i, o1 := range q1.Options {
+		if o1 != q2.Options[i] {
+			t.Errorf("AddressQuery Option at position %d mismatch", i)
+		}
+	}
+	CheckSubjectAddress(q1.SubjectAddr, q2.SubjectAddr, t)
+}
+
+func CheckSubjectAddress(a1, a2 *net.IPNet, t *testing.T) {
+	if a1.String() != a2.String() {
+		t.Error("SubjectAddr mismatch")
+	}
+}
+
+func CheckObjects(objs1, objs2 []rainslib.Object, t *testing.T) {
+	if len(objs1) != len(objs2) {
+		t.Error("Objects length mismatch")
+	}
+	for i, o1 := range objs1 {
+		o2 := objs2[i]
+		if o1.Type != o2.Type {
+			t.Errorf("Object Type mismatch at position %d", i)
+		}
+		switch o1.Type {
+		case rainslib.OTName:
+			n1 := o1.Value.(rainslib.NameObject)
+			n2 := o2.Value.(rainslib.NameObject)
+			if n1.Name != n2.Name {
+				t.Errorf("Object Value name Name mismatch at position %d", i)
+			}
+			if len(n1.Types) != len(n2.Types) {
+				t.Error("Object Value name types length mismatch")
+			}
+			for j, t1 := range n1.Types {
+				if t1 != n2.Types[j] {
+					t.Errorf("Object Value name type mismatch at byte %d of object %d", j, i)
+				}
+			}
+		case rainslib.OTIP6Addr:
+			if o1.Value.(string) != o2.Value.(string) {
+				t.Errorf("Object Value IP6 mismatch at position %d", i)
+			}
+		case rainslib.OTIP4Addr:
+			if o1.Value.(string) != o2.Value.(string) {
+				t.Errorf("Object Value IP4 mismatch at position %d", i)
+			}
+		case rainslib.OTRedirection:
+			if o1.Value.(string) != o2.Value.(string) {
+				t.Errorf("Object Value redirection mismatch at position %d", i)
+			}
+		case rainslib.OTDelegation:
+			CheckPublicKey(o1.Value.(rainslib.PublicKey), o2.Value.(rainslib.PublicKey), t)
+		case rainslib.OTNameset:
+			if o1.Value.(rainslib.NamesetExpression) != o2.Value.(rainslib.NamesetExpression) {
+				t.Errorf("Object Value nameSet mismatch at position %d", i)
+			}
+		case rainslib.OTCertInfo:
+			c1 := o1.Value.(rainslib.CertificateObject)
+			c2 := o2.Value.(rainslib.CertificateObject)
+			if c1.Type != c2.Type {
+				t.Errorf("Object Value CertificateInfo type mismatch at position %d", i)
+			}
+			if c1.HashAlgo != c2.HashAlgo {
+				t.Errorf("Object Value CertificateInfo HashAlgo mismatch at position %d", i)
+			}
+			if c1.Usage != c2.Usage {
+				t.Errorf("Object Value CertificateInfo Usage mismatch at position %d", i)
+			}
+			if len(c1.Data) != len(c2.Data) {
+				t.Errorf("Object Value CertificateInfo data length mismatch of object %d", i)
+			}
+			for j, b1 := range c1.Data {
+				if b1 != c2.Data[j] {
+					t.Errorf("Object Value CertificateInfo data mismatch at byte %d of object %d", j, i)
+				}
+			}
+		case rainslib.OTServiceInfo:
+			s1 := o1.Value.(rainslib.ServiceInfo)
+			s2 := o2.Value.(rainslib.ServiceInfo)
+			if s1.Name != s2.Name {
+				t.Errorf("Object Value service info name mismatch at position %d", i)
+			}
+			if s1.Port != s2.Port {
+				t.Errorf("Object Value service info Port mismatch at position %d", i)
+			}
+			if s1.Priority != s2.Priority {
+				t.Errorf("Object Value service info Priority mismatch at position %d", i)
+			}
+		case rainslib.OTRegistrar:
+			if o1.Value.(string) != o2.Value.(string) {
+				t.Errorf("Object Value registrar mismatch at position %d", i)
+			}
+		case rainslib.OTRegistrant:
+			if o1.Value.(string) != o2.Value.(string) {
+				t.Errorf("Object Value registrant mismatch at position %d", i)
+			}
+		case rainslib.OTInfraKey:
+			CheckPublicKey(o1.Value.(rainslib.PublicKey), o2.Value.(rainslib.PublicKey), t)
+		case rainslib.OTExtraKey:
+			CheckPublicKey(o1.Value.(rainslib.PublicKey), o2.Value.(rainslib.PublicKey), t)
+		}
+	}
+}
+
+func CheckPublicKey(p1, p2 rainslib.PublicKey, t *testing.T) {
+	if p1.KeySpace != p2.KeySpace {
+		t.Error("SubjectAddr KeySpace mismatch")
+	}
+	if p1.Type != p2.Type {
+		t.Error("SubjectAddr Type mismatch")
+	}
+	if p1.ValidSince != p2.ValidSince {
+		t.Error("SubjectAddr ValidSince mismatch")
+	}
+	if p1.ValidUntil != p2.ValidUntil {
+		t.Error("SubjectAddr ValidUntil mismatch")
+	}
+	switch p1 := p1.Key.(type) {
+	case rainslib.Ed25519PublicKey:
+		if p2, ok := p2.Key.(rainslib.Ed25519PublicKey); ok {
+			if p1 != p2 {
+				t.Errorf("publickey key mismatch p1=%v != %v=p2", p1, p2)
+			}
+		} else {
+			t.Errorf("publickey key type mismatch. Got Type:%T", p2)
+		}
+	default:
+		t.Errorf("Not yet supported. Got Type:%T", p1)
+	}
+}
